requirements/base.txt
# Core dependencies
django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
django-filter>=23.5
dj-database-url>=2.1
whitenoise>=6.6

# Supabase Storage
storage3>=0.9
supabase>=2.0
httpx>=0.24
